scraper/requirements.txt
# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Database
supabase>=2.3.0

# AI/ML
groq>=0.4.0

# Data Processing
python-dotenv>=1.0.0

# Utilities
dateparser>=1.2.0
python-dateutil>=2.8.2
